requirements.txt
numpy
deepface
fastapi
pydantic
uvicorn
python-multipart
python-dotenv
requests
aiofiles
pandas
gdown
tqdm
Pillow
# opencv-python
# tensorflow
# keras
# mtcnn>=0.1.0
# retina-face>=0.0.1
fire
Deprecated